+++ src/pagegen/pagegen.py
@@ -1,4 +1,4 @@
-from pagegen.utility import report_error, report_notice, get_site_conf_path, SITECONF, HOME, CONFROOT, TARGETDIR, HOOKDIR, CONTENTDIR, exec_script
+from pagegen.utility import report_error, report_notice, get_site_conf_path, SITECONF, HOME, CONFROOT, TARGETDIR, HOOKDIR, CONTENTDIR, exec_script, ASSETDIR, THEMEDIR
 from pagegen.site import site
 from os.path import expanduser, basename, join, isfile
 from os import getcwd, listdir, sep, chdir, X_OK, access
@@ -116,12 +116,11 @@ def serve_mode(site_conf_path, environment):
 	build_site(site_conf_path, environment, exclude_hooks, serve_base_url + ':' + serve_port, serve_mode=True)
 
 	watch_elements = [
-		'content',
-		'header_profiles',
-		'hooks',
-		'include',
-		'site.conf',
-		'templates'
+		CONTENTDIR,
+		HOOKDIR,
+		ASSETDIR,
+		SITECONF,
+		THEMEDIR
 	]
 
 	watch_elements_full_path = [site_dir + '/' + we for we in watch_elements]